restructure utils directory in preparation for additional Postman feature
