+++ src/utils/automation.ts
@@ -1,5 +1,5 @@
-import algoOne from "./utilityFunctions/tempAlgoOne";
-import algoTwo from "./utilityFunctions/tempAlgoTwo";
+import algoOne from "./buildPathObject";
+import algoTwo from "./buildSupertest";
 
 // This is the main Algorithm that will generate the superTest Code
 export default function( fileList:any ){
@@ -10,14 +10,14 @@ export default function( fileList:any ){
     // When the PathObject is sucessfully Built
     .then( pathObject => {
       // Create superTestCode and return it from the promise
-      console.log('PATH OBJECT:', pathObject)
+     
       const serverFile: string = pathObject[ pathObject.__serverFilePath__ ];
       const superTestCode = algoTwo(pathObject, serverFile);
       resolve(superTestCode)
     })
     // When there is an Error in the first Algorithm
     .catch(pathObject => {
-      console.log('ERROR OCCURRED: ', pathObject)
+      
       // Attempt to create the superTest Code and return the result from the promise
       const serverFile: string = pathObject[ pathObject.__serverFilePath__ ];
       const superTestCode = algoTwo(pathObject, serverFile);
