+++ src/utils/buildPostman.ts
@@ -1,0 +1,7 @@
+// buildPathObject result set in state
+// pull path object from state 
+// pass to Postman or Supertest
+
+// buildPathObject(fileList)
+//  .then( pathObject => set state with it here)
+// buildSuperTest( axonState.pathObject)
